Complete task
I've allowed the playGame function to run through 5 rounds and tally
results to output a winner.

I've also rectified some code in the playRound so that when someone
enters an invalid choice, it doesn't just tell them it was invalid, but
asks them for the choice again to get a valid result. It's also case
insensitive.

Unfortunately as I've not been taught loops yet I had to just copy and
paste the playRound if statements 5 times but the exercise says that's
ok.