Complete task and expand
I have finished editing the code so that it tracks
game winners and has a 'first to five' system in
place. It displays the winner each round and game
winner.

I've started to add rudimentary animations. I have
made the elements of the page fade in, staggered.
There are one or two tweaks, and then some more
hefty animations I'd like to try and add before I
am done.